// src/model/User.ts
export class User {
         userId?: any;
         mail: any;
         passWord: any;
         userName: any;
         photoUser?: any;
         dateDeNaissance?: any;
         ville: any;
         codePostal: any;
         adresse: any;
         userSex: any;
         experience1?: any;
         experience2?: any;
         experience3?: any;
         competence1: any;
         competence2?: any;
         competence3?: any;
         formation1?: any;
         formation2?: any;
         formation3?: any;
         CVpdf?: any;
         telephone: any;
         Linkedin?: any;
         metier: any;
         Phrase_d_accroche: any;
       }